minor change in models.py

+++ blog_project/models.py
@@ -31,10 +31,11 @@ class User(db.Model, UserMixin):
 
     def unlike(self, post):
         if self.has_liked_post(post):
-            PostLike.query.filter_by(
-                user_id=self.id,
-                post_id=post.id)\
-                .delete()
+            PostLike.query.filter_by(user_id=self.id, post_id=post.id).delete()
+
+    def has_liked(self, post):
+        return PostLike.query.filter(PostLike.user_id == self.id,
+                                     PostLike.post_id == post.id).count() > 0
 
 
 class GoogleUser(User):
